refactor: step2 중복 container 코드 제거, delegate 파일 분리
