+++ Notes/Protocol/NoteSelectionDelegate.swift
@@ -1,0 +1,12 @@
+//
+//  NoteSelectionDelegate.swift
+//  Notes
+//
+//  Created by kakao on 2022/01/21.
+//
+
+import Foundation
+
+protocol NoteSelectionDelegate: AnyObject {
+    func noteSelected(_ note: Note)
+}
